+++ bin/serve.ts
@@ -1,5 +1,0 @@
-import { Buchta } from "../src/buchta";
-
-const buchta = new Buchta();
-
-buchta.run();