+++ plugins/svelte.ts
@@ -245,6 +245,24 @@ ${code}
             opts.set("livereload", {pretty: "Enable livereload?", type: BuchtaQuestionType.YES_OR_NO});
 
             this.setTemplateOptions("svelte", opts);
+
+            this.setPluginTemplate("svelte", {
+                filename: "svelte.html",
+                content: `
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+</head>
+<body>
+<!-- html -->
+</body>
+<!-- code -->
+</html>
+`
+            });
             
             this.setTemplateCallback("svelte", (opts: Map<string, BuchtaProjectOption>) => {
                 const configTemplate = `